[Refactor] 코디네이터 패턴 적용 완료

+++ Projects/Features/MainFeature/Interface/Sources/MainPresentable.swift
@@ -12,7 +12,7 @@ import Core
 
 public protocol MainViewControllable: ViewControllable { }
 public protocol MainCoordinatable {
-    var onWeatherCellTap: (() -> Void)? { get set }
+    var onWeatherCellTap: ((Int) -> Void)? { get set }
 }
 public typealias MainViewModelType = ViewModelType & MainCoordinatable
 public typealias MainPresentable = (vc: MainViewControllable, vm: any MainViewModelType)
